Delete comments and votes on story deletion

+++ internal/vote/vote.go
@@ -50,10 +50,9 @@ func voteUpdate(ctx context.Context, voter user.User, storyID uuid.UUID, directi
 		}
 	}
 
-	// Increment vote for the user.
+	// Save vote for the user.
 	{
 		var v int64
-		// TODO(grafov) Better use upsert here but I have deadline soon. Replace it later.
 		const q = `SELECT vote FROM vote WHERE story_id = $1 AND account_id = $2`
 		if err = tx.QueryRowxContext(ctx, q, storyID, voter.ID).Scan(&v); err != nil && err != sql.ErrNoRows {
 			l.Log("err", err, "sql", q, "desc", "vote load failed")
@@ -66,23 +65,19 @@ func voteUpdate(ctx context.Context, voter user.User, storyID uuid.UUID, directi
 				l.Log("err", err, "sql", q, "desc", "vote insert failed")
 				return []Vote{}, errInternal
 			}
+			// Update story' score counter.
+			const q2 = `UPDATE story SET score = score + $2 WHERE id = $1`
+			if _, err = tx.ExecContext(ctx, q2, storyID, direction); err != nil {
+				l.Log("err", err, "sql", q2, "desc", "vote load failed")
+				return []Vote{}, errInternal
+			}
 		}
-		// This case handles existing scores: reset or increment/decrement.
-		if err != sql.ErrNoRows {
-			if direction != 0 {
-				// Increment or decrement first vote for the user.
-				const q = `UPDATE vote SET vote = vote + $3 WHERE story_id = $1 AND account_id = $2`
-				if _, err = tx.ExecContext(ctx, q, storyID, voter.ID, direction); err != nil {
-					l.Log("err", err, "sql", q, "desc", "vote update failed")
-					return []Vote{}, errInternal
-				}
-			} else {
-				// Reset the score for the user (direction = 0).
-				const q = `DELETE FROM vote WHERE story_id = $1 AND account_id = $2`
-				if _, err = tx.ExecContext(ctx, q, storyID, voter.ID); err != nil {
-					l.Log("err", err, "sql", q, "desc", "vote update failed")
-					return []Vote{}, errInternal
-				}
+		// This case handles score reset. Increment/decrement skiped here.
+		if err != sql.ErrNoRows && direction == 0 {
+			const q = `DELETE FROM vote WHERE story_id = $1 AND account_id = $2`
+			if _, err = tx.ExecContext(ctx, q, storyID, voter.ID); err != nil {
+				l.Log("err", err, "sql", q, "desc", "vote update failed")
+				return []Vote{}, errInternal
 			}
 		}
 	}
